Update proposal
Hi Gavin,

I'm trying to implement your solution in our environment. But when I change the boot_type to "LEGACY" Nutanix is unhappy:
 "requested_state_transition": "kNone",
  "response": {
    "error_code": 10,
    "error_detail": "A disk_address, mac_address, boot device order or must be specified for the non UEFI boot configuration"
  }

Hence I want to modify the template file to support this. The above is just an impression and not working yet. I'll let you know as soon as I  figure it out.